### 2021-03-08 #3034 @vissarion
:wave: @GISerDaishaoqing, @becarioprecario, @fpl, @fabian-s, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
