### 2021-03-08 #3092 @whedon
[REVIEW]: bleiglas: An R package for interpolation and visualisation of spatiotemporal data with 3D tessellation
**Submitting author:** @nevrome (<a href="http://orcid.org/0000-0003-3448-5715">Clemens Schmid</a>)
**Repository:** <a href="https://github.com/nevrome/bleiglas" target ="_blank">https://github.com/nevrome/bleiglas</a>
**Version:** 1.0.0
**Reviewer:** @corybrunson , @fabian-s
**Archive:** <a href="https://doi.org/10.17605/OSF.IO/AGMD6" target="_blank">10.17605/OSF.IO/AGMD6</a>
