### 2021-03-08 #3034 @becarioprecario
Hi,

Thanks but I am not available at the moment.

Best,

Virgilio

El 8 mar 2021, a las 9:35, Vissarion Fisikopoulos <notifications@github.com<mailto:notifications@github.com>> escribió:

👋 @GISerDaiShaoqing<https://github.com/GISerDaiShaoqing>, @becarioprecario<https://github.com/becarioprecario>, @fpl<https://github.com/fpl>, @fabian-s<https://github.com/fabian-s>, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
