### 2021-03-08 #3092 @fabian-s
- [x] **Installation:** Does installation proceed as outlined in the documentation? (But see bleiglas/#7)
