### 2021-03-15 nevrome/bleiglas#12 @corybrunson
package c14bazAAR has been removed from CRAN
The README illustrates **bleiglas** functionality using the data set `c14` installed with the package **c14bazAAR**. However, the latter package has been removed from CRAN. (Check [the CRAN page](https://cran.r-project.org/package=c14bazAAR) to be sure.) Depending on that package's permissions, the data set could be installed with **bleiglas** instead; or, another data set could be used for illustration. But i believe it is inadvisable (anyway, i would advise against it!) to rely on a package that has been removed from CRAN, perhaps unless it is under active development with plans to restore it.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3092)._
