### 2021-03-15 nevrome/bleiglas#13 @corybrunson
naming and installing vignette
When a package is installed from GitHub using **remotes**, the option `build_vignettes = TRUE` must be made explicit so that the user can view the vignettes. So i recommend including this option in the README installation instructions. (This might not become a moot point when the package is submitted to CRAN, since some users may still want to install development versions from GitHub.)

While on the subject, though, vignettes are commonly accessed via `vignette("<name of vignette>")`, without specifying `package = "<package>"`. So it would be beneficial to give every vignette a specific name having to do with the functionality it illustrates, possibly even the package name (especially if the package only includes one vignette).

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3092)._
