### 2021-03-15 nevrome/bleiglas#14 @corybrunson
package dependencies in scripts referenced by the vignette
The vignette refers to, but does not evaluate, scripts installed in "inst/workflow_example", which seem to depend not only on the **rnaturalearth** package but also on packages including **rgdal** that are required by specific steps. I'm not sure what the exact protocol is for such a setup (unless i've missed something, [the R extensions manual](https://cran.r-project.org/doc/manuals/R-exts.html#Package-Dependencies) is silent on this), but should these packages at least be listed as dependencies (presumably under `Suggests:`) in DESCRIPTION? This is indeed a question, not a recommendation, since the vignette can be generated (and **bleiglas** can be installed) without installing these.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3092)._
