### 2021-03-22 #3092 @corybrunson
Update on my review: I've test-run the software and will review the documentation and paper this week. I've been creating issues on the repo and will post here when i finish—with any additional comments not made in the issues.
