### 2021-03-24 #3092 @corybrunson
@vissarion should the two "statement of need" requirements in the reviewer checklist be interchanged? The one in the "software paper" section reads like it refers to the documentation, and vice-versa. And could you confirm that a statement of need is indeed expected in both places? Thank you!
