### 2021-03-24 nevrome/bleiglas#15 @corybrunson
link to data.frame in documentation
data.tables are used throughout the package and mentioned throughout the documentation, but these mentions never link to the data.table package itself, where the central object is documented at `data.table::data.table`. Since the package is intended for potentially quite general use, i think it would be helpful to include at least a link to this object, so that those only used to data.frames will be able to more efficiently learn about these objects and their methods.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3092)._
