### 2021-03-24 #3092 @corybrunson
@vissarion I have completed my initial review. @nevrome this package was well-organized and straightforward to evaluate, and the paper was concise and clearly written, for all of which i thank you! All of the issues i identified have been mentioned in issues, either my own or (often) those by @fabian-s before me. I'll be available to answer follow-up questions and will complete my review checklist as the issues are addressed.
