### 2021-03-24 3fadd6d
improved install instructions for voro++ - if merged this closes #7
