+++ README.Rmd
@@ -32,7 +32,7 @@ bleiglas is an R package that provides functions for 3D tessellation with [Voro+
 ## Documentation
 
 1. This README (see Quickstart guide below) describes a basic workflow with code and explains some of my thought process when writing this package
-2. A [JOSS paper](https://github.com/nevrome/bleiglas/blob/master/paper/paper.md) (not submitted yet) which gives some background, introduces the core functions from a more technical point of view and presents an example application
+2. A [JOSS paper](https://github.com/nevrome/bleiglas/blob/master/paper/paper.md) (in Review) which gives some background, introduces the core functions from a more technical point of view and presents an example application
 3. A [vignette](https://github.com/nevrome/bleiglas/blob/master/vignettes/complete_example.Rmd) which contains all the code necessary to reproduce the example application in said JOSS paper
 
 ## Installation 
@@ -44,7 +44,7 @@ if(!require('remotes')) install.packages('remotes')
 remotes::install_github("nevrome/bleiglas")
 ```
 
-For the main function `tessellate` you also have to [install the Voro++ software](http://math.lbl.gov/voro++/download/). For Linux users: The package is already available in all major software repositories.
+For the main function `tessellate` you also have to [install the Voro++ software](http://math.lbl.gov/voro++/download/). The package is already available in all major Linux software repositories (on Debian/Ubuntu you can simply run `sudo apt-get install voro++`.). MacOS users should be able to install it via homebrew (`brew install voro++`).
 
 ## Quickstart
 
