### 2021-03-24 e2a5c65
added some contributing guidelines (initially taken from rOpenSci?) #9
