### 2021-03-24 2afd275
added a sentence to the README linking to the contributing guide #9
