+++ README.Rmd
@@ -62,7 +62,9 @@ I decided to use Dirk Seidenstickers [*Archives des datations radiocarbone d'Afr
 I selected dates from Cameroon between 1000 and 3000 uncalibrated BP and projected them into a worldwide cylindrical reference system (epsg [4088](https://epsg.io/4088)). As Cameroon is close to the equator this projection should represent distances, angles and areas sufficiently correct for this example exercise. As a minor pre-processing step, I here also remove samples with equal position in all three dimensions for the tessellation.
 
 ```{r, message=FALSE}
-# download raw data
+# download raw data with the data access package c14bazAAR
+# c14bazAAR can be installed with
+# install.packages("c14bazAAR", repos = c(ropensci = "https://ropensci.r-universe.dev"))
 c14_cmr <- c14bazAAR::get_c14data("adrac") %>% 
   # filter data
   dplyr::filter(!is.na(lat) & !is.na(lon), c14age > 1000, c14age < 3000, country == "CMR") 
