### 2021-03-24 nevrome/bleiglas#12 @nevrome
I'm the maintainer of c14bazAAR and it was removed from CRAN because of an issue in a dependency, not the package itself. c14bazAAR is still actively developed and part of the rOpenSci initiative, but I'm not planning to resubmit to CRAN at the moment.

I hope 8855d24 somehow mitigates your concern.
