### 2021-03-24 c673dac
renamed vignette and improved the related install instructions in the README #13
