### 2021-03-24 fb65fad
added explicit dependency installation instructions at the beginning of the vignette
