+++ vignettes/bleiglas_bronze_age_burial_rite_distribution_case_study.Rmd
@@ -27,6 +27,20 @@ code_files <- basename(code_files_full_path)
 sourcecode <- lapply(code_files_full_path, function(x) { paste(readLines(x), collapse="\n") })
 ```
 
+Running the code in this vignette requires some additional packages. You can install them with
+
+``` {r eval=FALSE}
+install.packages(
+  c("Bchron", "bleiglas", "c14bazAAR", "data.table", "dplyr", 
+    "ggplot2", "magrittr", "pbapply", "purrr", "raster", 
+    "rnaturalearth", "scatterplot3d", "sf", "tibble"),
+  repos = c(
+    CRAN = "https://cloud.r-project.org", 
+    ropensci = "https://ropensci.r-universe.dev"
+  )
+)
+```
+
 ## `r code_files[1]`
 
 This first script downloads and prepares a set of spatial data objects which will later be used for plotting. The research area for this example was arbitrarily defined as a rectangle covering the most dense data accumulations in the relevant subset of the Radon-B database.
