### 2021-03-24 nevrome/bleiglas#14 @nevrome
The setup is indeed unusual. Probably the Quickstart guide in the README would be more typical for a vignette. But then again vignettes are used for many different purposes (some times even whole research papers) and I think this extensive compilation of commented code could be useful for someone who really wants to use this package to the fullest.

I would rather not have the heavy dependencies for this specific case study in the package DESCRIPTION though. Instead I added explicit installation instructions at the beginning of the vignette in 8855d24. Do you think this makes it reasonably accessible?
