### 2021-03-24 3218f7c
added some links to the data.table documentation #15
