+++ man/cut_polygons.Rd
@@ -7,7 +7,7 @@
 cut_polygons(x, cuts)
 }
 \arguments{
-\item{x}{data.table with output of voro++ as produced with
+\item{x}{\link[data.table]{data.table} with output of voro++ as produced with
 \link{tessellate} and then \link{read_polygon_edges}}
 
 \item{cuts}{numeric vector with z-axis coordinates where cuts should be applied}
