### 2021-03-24 9155a80
replaced sapply with the type safe vapply #11
