### 2021-03-24 nevrome/bleiglas#12 @corybrunson
Yes, this is quite satisfactory! Thank you.
