### 2021-03-24 nevrome/bleiglas#15 @corybrunson
No, this is great! Having both the detailed sentence in the documentation for the primary function and the hyperlink in that for the auxiliary functions makes sense.
