+++ DESCRIPTION
@@ -2,7 +2,6 @@ Package: bleiglas
 Title: Spatiotemporal Data Interpolation and Visualisation based on 3D Tessellation with Voro++
 Description: 3D tessellation with the Voro++ software library and subsequent horizontal cutting of the resulting polygons for 2D representation. Applicable to all sorts of 3D point clouds, but optimized for spatiotemporal data.
 Version: 0.1.2
-Date: 2020-07-22
 Authors@R: 
     c(
       person(given = "Clemens",
