### 2021-03-25 babf409
defined a Statement of need section (#8) in the JOSS paper and added a section about tools for voronoi calculation in R (#10)
