+++ paper/paper.bib
@@ -119,3 +119,31 @@
   author = {Jutta Kneisel and Martin Hinz and Christoph Rinne},
   year = {2013},
 }
+@Manual{Turner2021,
+  title = {deldir: Delaunay Triangulation and Dirichlet (Voronoi) Tessellation},
+  author = {Rolf Turner},
+  year = {2021},
+  note = {R package version 0.2-10},
+  url = {https://CRAN.R-project.org/package=deldir},
+}
+@Manual{Gebhardt2020,
+  title = {tripack: Triangulation of Irregularly Spaced Data},
+  author = {Albrecht Gebhardt and R. J. Renka and Stephen Eglen and Sergei Zuyev and Denis White},
+  year = {2020},
+  note = {R package version 1.3-9.1},
+  url = {https://CRAN.R-project.org/package=tripack},
+}
+@Manual{Hijmans2020,
+  title = {dismo: Species Distribution Modeling},
+  author = {Robert J. Hijmans and Steven Phillips and John Leathwick and Jane Elith},
+  year = {2020},
+  note = {R package version 1.3-3},
+  url = {https://CRAN.R-project.org/package=dismo},
+}
+@Manual{Garrett2021,
+  title = {ggvoronoi: Voronoi Diagrams and Heatmaps with 'ggplot2'},
+  author = {Robert C. Garrett and Austin Nar and Thomas J. Fisher},
+  year = {2021},
+  note = {R package version 0.8.4},
+  url = {https://CRAN.R-project.org/package=ggvoronoi},
+}
