### 2021-03-26 #3092 @vissarion
> @vissarion should the two "statement of need" requirements in the reviewer checklist be interchanged? The one in the "software paper" section reads like it refers to the documentation, and vice-versa. And could you confirm that a statement of need is indeed expected in both places? Thank you!

It makes sense to ask for a statement of need both in the documentation and in the JOSS paper but I agree that the text in the checklist should be changed. Of course this would not affect the current review process. I propose to threat each "statement of need" check boxes separately, i.e. one for documentation and one for the paper. Thanks for mentioning this @corybrunson.
