### 2021-03-26 ee6de83
added a contact email to the Code of Conduct
