### 2021-03-26 nevrome/bleiglas#13 @corybrunson
The new title is very helpful! But its transliteration to the vignette name (what the user has to provide to `vignette()`) might be cumbersome. Natural shorter names might include "bleiglas", "burial-rites", or "spatiotemporal-tessellation", depending on what you see as the role of the vignette in the package (primary/introductory tutorial, use case, or illustration of specific functionality).
