+++ README.Rmd
@@ -33,7 +33,7 @@ bleiglas is an R package that provides functions for 3D tessellation with [Voro+
 
 1. This README (see Quickstart guide below) describes a basic workflow with code and explains some of my thought process when writing this package
 2. A [JOSS paper](https://github.com/nevrome/bleiglas/blob/master/paper/paper.md) (in Review) which gives some background, introduces the core functions from a more technical point of view and presents an example application
-3. A (rather technical) vignette which contains all the code necessary to reproduce the example application in said JOSS paper. When bleiglas is installed you can open the vignette in R with `vignette("bleiglas_bronze_age_burial_rite_distribution_case_study")`
+3. A (rather technical) vignette which contains all the code necessary to reproduce the example application in said JOSS paper. When bleiglas is installed you can open the vignette in R with `vignette("bleiglas_case_study")`
 
 If you have questions beyond this documentation feel free to open an [issue](https://github.com/nevrome/bleiglas/issues) here on Github. Please also see our [contributing guide](CONTRIBUTING.md).
 
