### 2021-03-26 nevrome/bleiglas#13 @nevrome
I went with `bleiglas_case_study` now. I think that's reasonably short and informative.
