### 2021-03-26 01371f0
better handling of the citation in the README - only preliminary though
