### 2021-03-26 512f68a
better package description in the DESCRIPTION file
