### 2021-03-26 f629656
more extensive introduction in the README
