### 2021-03-29 nevrome/bleiglas#17 @vissarion
JOSS paper bibliography 
There are some minor issues in the JOSS bibliography
- In reference https://github.com/nevrome/bleiglas/blob/master/paper/paper.bib#L143 it is better to use the DOI of the package (10.21105/joss.01096) instead of the cran link, see https://github.com/openjournals/joss-reviews/issues/3092#issuecomment-809160528
-  This ref https://github.com/nevrome/bleiglas/blob/master/paper/paper.bib#L116 could have a clickable link to https://radon-b.ufg.uni-kiel.de/
- https://github.com/nevrome/bleiglas/blob/master/paper/paper.bib#L122 voronoi and dirichlet are names and should be capital.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3092).
