### 2021-03-29 #3092 @vissarion
@nevrome there are some minor issues in your bibliography, see https://github.com/nevrome/bleiglas/issues/17
