+++ paper/paper.md
@@ -31,7 +31,7 @@ Voronoi tessellation has useful applications in all kinds of scientific contexts
 
 The ``bleiglas`` R package serves as an R interface to the Voro++ command line tool. It adds a number of utility functions for particular data manipulation applications, including but not limited to automatic 2D cutting of the 3D Voro++ output for subsequent mapping and grid sampling for position and value uncertainty mitigation. The relevant workflows are explained below. Although we wrote this package for our own needs in archaeology and archaeogenetics, the code is by no means restricted to data from these fields, just as Voronoi tessellation is a generic, subject-agnostic method with a huge range of use-cases.
 
-Voronoi tessellation is implemented in a many R packages, maybe most prominently in the deldir (*Delaunay Triangulation and Dirichlet (Voronoi) Tessellation*) [@Turner2021] and the tripack (*Triangulation of Irregularly Spaced Data*) [@Gebhardt2020] packages, which were specifically designed for this application. ggvoronoi (*Voronoi Diagrams and Heatmaps with 'ggplot2'*) [@Garrett2021] and `dismo::voronoi()` [@Hijmans2020] build on deldir. Another implementation directly integrated with the spatial `sf` [@Pebesma2018] format is available with `sf::st_voronoi()`. All of these packages and functions focus on 2D data and to our knowledge none offer a comparable toolset to handle 3D Voronoi diagrams as introduced with ``bleiglas``.
+Voronoi tessellation is implemented in a many R packages, maybe most prominently in the deldir (*Delaunay Triangulation and Dirichlet (Voronoi) Tessellation*) [@Turner2021] and the tripack (*Triangulation of Irregularly Spaced Data*) [@Gebhardt2020] packages, which were specifically designed for this application. ggvoronoi (*Voronoi Diagrams and Heatmaps with 'ggplot2'*) [@Garrett2018] and `dismo::voronoi()` [@Hijmans2020] build on deldir. Another implementation directly integrated with the spatial `sf` [@Pebesma2018] format is available with `sf::st_voronoi()`. All of these packages and functions focus on 2D data and to our knowledge none offer a comparable toolset to handle 3D Voronoi diagrams as introduced with ``bleiglas``.
 
 # Core functionality
 
