+++ paper/paper.bib
@@ -140,10 +140,15 @@
   note = {R package version 1.3-3},
   url = {https://CRAN.R-project.org/package=dismo},
 }
-@Manual{Garrett2021,
-  title = {ggvoronoi: Voronoi Diagrams and Heatmaps with 'ggplot2'},
-  author = {Robert C. Garrett and Austin Nar and Thomas J. Fisher},
-  year = {2021},
-  note = {R package version 0.8.4},
-  url = {https://CRAN.R-project.org/package=ggvoronoi},
+@article{Garrett2018,
+  doi = {10.21105/joss.01096},
+  url = {https://doi.org/10.21105/joss.01096},
+  year = {2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {32},
+  pages = {1096},
+  author = {Robert C. Garrett and Austin Nar and Thomas J. Fisher and Karsten Maurer},
+  title = {ggvoronoi: Voronoi Diagrams and Heatmaps with ggplot2},
+  journal = {Journal of Open Source Software}
 }
