+++ paper/paper.md
@@ -45,7 +45,7 @@ A prerequisite for performing tessellation in three dimensions is the normalisat
 
 # Example: Burial rite distributions in Bronze Age Europe
 
-One strength of ``bleiglas`` is visualisation of spatiotemporal data. Here we show an example of Bronze Age burial rites as measured on radiocarbon dates from burials in Central, Northern and Northwestern Europe between 2200 and 800 calBC. Information about source data (taken from the RADON-B database [@kneiselRadonB2013]), data preparation and meaning are presented in @Schmid2019. The "Complete example" vignette (`vignette("complete_example", package = "bleiglas")`) contains the code to reproduce the following figures.
+One strength of ``bleiglas`` is visualisation of spatiotemporal data. Here we show an example of Bronze Age burial rites as measured on radiocarbon dates from burials in Central, Northern and Northwestern Europe between 2200 and 800 calBC. Information about source data (taken from the [RADON-B database](https://radon-b.ufg.uni-kiel.de) [@kneiselRadonB2013]), data preparation and meaning are presented in @Schmid2019. The "Complete example" vignette (`vignette("complete_example", package = "bleiglas")`) contains the code to reproduce the following figures.
 
 Bronze Age burials can be classified by two main aspects: inhumation vs. cremation (*burial type*) and flat grave vs. burial mound (*burial construction*). \autoref{fig:plot_map} is a map of burials through time for which we have some information about these variables. Each grave has a position in space (coordinates) and in time (median calibrated radiocarbon age). For \autoref{fig:plot_3D} and \autoref{fig:plot_bleiglas} we only look at the *burial type* aspect. The burials are distributed in a three dimensional, spatiotemporal space and therefore can be subjected to Voronoi tessellation with Voro++. As detailed above, the outcome depends on the relative scaling of the input dimensions - for this example we choose $1\text{kilometer}=1\text{year}$, informed by some intuition about the range of human movement through time.
 
