### 2021-03-29 7307411
added case protection for some title elements in the bibtex file #17
