+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{Rycroft2009-rp,
-  title    = "{VORO++}: a three-dimensional voronoi cell library in C++",
+  title    = {{Voro++}: a three-dimensional {Voronoi} cell library in {C++}},
   author   = "Rycroft, Chris H",
   journal  = "Chaos",
   volume   =  19,
@@ -40,8 +40,8 @@
   doi      = {10.1016/j.compenvurbsys.2019.101357},
 }
 @ARTICLE{Paranjape2020-sg,
-  title     = "Voronoi volume function: a new probe of cosmology and galaxy
-               evolution",
+  title     = {{Voronoi} volume function: a new probe of cosmology and galaxy
+               evolution},
   author    = "Paranjape, Aseem and Alam, Shadab",
   journal   = "Monthly Notices of the Royal Astronomical Society",
   publisher = "Oxford Academic",
@@ -65,14 +65,14 @@
   doi       = {10.1007/978-3-319-29538-1},
 }
 @Manual{Dowle2019,
-  title = {data.table: Extension of `data.frame`},
+  title = {{data.table}: Extension of {data.frame}},
   author = {Matt Dowle and Arun Srinivasan},
   year = {2019},
   note = {R package version 1.12.8},
   url = {https://CRAN.R-project.org/package=data.table},
 }
 @Article{Eddelbuettel2017,
-  title = {Extending R with C++: A Brief Introduction to Rcpp},
+  title = {Extending {R} with {C++}: A Brief Introduction to {Rcpp}},
   author = {Dirk Eddelbuettel and James Joseph Balamuta},
   journal = {PeerJ Preprints},
   year = {2017},
@@ -85,7 +85,7 @@
 }
 @Article{Pebesma2018,
   author = {Edzer Pebesma},
-  title = {{Simple Features for R: Standardized Support for Spatial Vector Data}},
+  title = {{Simple Features for {R}: Standardized Support for Spatial Vector Data}},
   year = {2018},
   journal = {{The R Journal}},
   doi = {10.32614/RJ-2018-009},
@@ -95,15 +95,15 @@
   number = {1},
 }
 @Manual{Schmid2017,
-  title = {recexcavAAR: 3D Reconstruction of Archaeological Excavations},
+  title = {{recexcavAAR}: 3D Reconstruction of Archaeological Excavations},
   author = {Clemens Schmid and Benjamin Serbe},
   year = {2017},
   note = {R package version 0.3.0},
   url = {https://CRAN.R-project.org/package=recexcavAAR},
 }
 @ARTICLE{Schmid2019,
-  title     = "Evaluating Cultural Transmission in Bronze Age burial rites of
-               Central, Northern and Northwestern Europe using radiocarbon data",
+  title     = {Evaluating Cultural Transmission in Bronze Age burial rites of
+               {Central, Northern and Northwestern Europe} using radiocarbon data},
   author    = "Schmid, Clemens",
   journal   = "Adaptive Behavior",
   publisher = "SAGE Publications Ltd STM",
@@ -120,21 +120,21 @@
   year = {2013},
 }
 @Manual{Turner2021,
-  title = {deldir: Delaunay Triangulation and Dirichlet (Voronoi) Tessellation},
+  title = {deldir: {Delaunay} Triangulation and {Dirichlet} ({Voronoi}) Tessellation},
   author = {Rolf Turner},
   year = {2021},
   note = {R package version 0.2-10},
   url = {https://CRAN.R-project.org/package=deldir},
 }
 @Manual{Gebhardt2020,
-  title = {tripack: Triangulation of Irregularly Spaced Data},
[…]
