### 2021-03-29 a919f04
added the geos package to the list of R packages with Voronoi implementations
