+++ paper/paper.bib
@@ -152,3 +152,16 @@
   title = {{ggvoronoi}: {Voronoi} Diagrams and Heatmaps with ggplot2},
   journal = {Journal of Open Source Software}
 }
+@Manual{Dunnington2021,
+  title = {{geos}: Open Source Geometry Engine {('GEOS') R API}},
+  author = {Dewey Dunnington and Edzer Pebesma},
+  year = {2021},
+  note = {R package version 0.0.2},
+  url = {https://CRAN.R-project.org/package=geos},
+}
+@Misc{geos2021,
+  title = {{GEOS} - Geometry Engine, Open Source},
+  howpublished = {https://trac.osgeo.org/geos.},
+  author = {{GEOS Development Team}},
+  year = {2021},
+}
