### 2021-03-29 80018a0
vignette name was changed during the JOSS review
