### 2021-03-29 f4b916b
added human-readable name of map projection
