+++ paper/paper.md
@@ -55,9 +55,9 @@ For the final \autoref{fig:plot_prediction_grid} we applied temporal resampling 
 
 ![Graves in the research area (rectangular frame) dating between 2200 and 800 calBC as extracted from the [Radon-B database](radon-b.ufg.uni-kiel.de). The classes of the variable burial type are distinguished by colour, the ones of burial construction by shape. The map projection is EPSG:102013 (Europe Albers Equal Area Conic) and the base layer data is taken from the [Natural Earth project](https://www.naturalearthdata.com/).\label{fig:plot_map}](03_map_plot.jpeg)
 
-![Graves in 3D space defined by two spatial (x and y in km) and a temporal (z in years calBC) dimension with Voronoi polygons constructed by Voro++. Each red dot represents one grave with known burial type, the fine black lines the edges of the result polygons, the rectangular wireframe box the research area now in space and time.\label{fig:plot_3D}](05_3D_plot.jpeg)
+![Graves in 3D space defined by two spatial (x and y in km) and a temporal (z in years calBC) dimension with Voronoi polygons constructed by Voro++. Each red dot represents one grave with known burial type, the fine black lines the edges of the result polygons and the rectangular wireframe box the research area now in space and time.\label{fig:plot_3D}](05_3D_plot.jpeg)
 
-![*bleiglas plot*. Map matrix of 2D cuts through 3D Voronoi polygons as presented in \autoref{fig:plot_3D}. Each subplot shows one 200 years timeslice between 2200 and 800 calBC. As each 2D polygon belongs to one input burial and data density in some areas and time periods is very low, some graves are represented in multiple subplots. Color coding and map background is as in \autoref{fig:plot_map}.\label{fig:plot_bleiglas}](06_bleiglas_plot.jpeg)
+![*bleiglas plot*. Map matrix of 2D cuts through 3D Voronoi polygons as presented in \autoref{fig:plot_3D}. Each subplot shows one timeslice between 2200 and 800 calBC. As each 2D polygon belongs to one input grave and data density in some areas and time periods is very low, some graves are represented in multiple subplots. Color coding and map background is as in \autoref{fig:plot_map}.\label{fig:plot_bleiglas}](06_bleiglas_plot.jpeg)
 
 ![Temporal resampling version of the bleiglas map matrix. 30 resampling runs and a spatial resolution of 100*100 cells in the research area shape bounding box. Color coding and map background again as in \autoref{fig:plot_map}.\label{fig:plot_prediction_grid}](07_prediction_grid_plot.jpeg)
 
