### 2021-03-29 #3092 @nevrome
@vissarion Thanks for the suggestions!

I created a PR to address these small, final proofreading changes: https://github.com/nevrome/bleiglas/pull/18

I also wait for final approval by @stschiff, which might take until after the Easter holidays.
