+++ paper/paper.md
@@ -31,11 +31,11 @@ Voronoi tessellation has useful applications in all kinds of scientific contexts
 
 The ``bleiglas`` R package serves as an R interface to the Voro++ command line tool. It adds a number of utility functions for particular data manipulation applications, including but not limited to automatic 2D cutting of the 3D Voro++ output for subsequent mapping and grid sampling for position and value uncertainty mitigation. The relevant workflows are explained below. Although we wrote this package for our own needs in archaeology and archaeogenetics, the code is by no means restricted to data from these fields, just as Voronoi tessellation is a generic, subject-agnostic method with a huge range of use-cases.
 
-Voronoi tessellation is implemented in a many R packages, maybe most prominently in the deldir (*Delaunay Triangulation and Dirichlet (Voronoi) Tessellation*) [@Turner2021] and the tripack (*Triangulation of Irregularly Spaced Data*) [@Gebhardt2020] packages, which were specifically designed for this application. ggvoronoi (*Voronoi Diagrams and Heatmaps with 'ggplot2'*) [@Garrett2018] and `dismo::voronoi()` [@Hijmans2020] build on deldir. Other implementations like `sf::st_voronoi()` [@Pebesma2018] or `geos::geos_voronoi_polygons()` [@Dunnington2021] rely on the [GEOS library](https://trac.osgeo.org/geos) [@geos2021]. All of these packages and functions focus on 2D data and to our knowledge none offer a comparable toolset to handle 3D Voronoi diagrams as introduced with ``bleiglas``.
+Voronoi tessellation is implemented in many R packages, maybe most prominently in the deldir (*Delaunay Triangulation and Dirichlet (Voronoi) Tessellation*) [@Turner2021] and the tripack (*Triangulation of Irregularly Spaced Data*) [@Gebhardt2020] packages, which were specifically designed for this application. ggvoronoi (*Voronoi Diagrams and Heatmaps with 'ggplot2'*) [@Garrett2018] and `dismo::voronoi()` [@Hijmans2020] build on deldir. Other implementations like `sf::st_voronoi()` [@Pebesma2018] or `geos::geos_voronoi_polygons()` [@Dunnington2021] rely on the [GEOS library](https://trac.osgeo.org/geos) [@geos2021]. All of these packages and functions focus on 2D data and to our knowledge none offer a comparable toolset to handle 3D Voronoi diagrams as introduced with ``bleiglas``.
 
 # Core functionality
 
-``bleiglas`` provides the `bleiglas::tessellate()` function which is a a command line utility wrapper for Voro++. It requires Voro++ to be installed locally. `tessellate()` takes input points in the form of a `data.frame` with an integer ID and three numeric coordinate columns. Additional Voro++ [options](http://math.lbl.gov/voro++/doc/cmd.html) can be set with a character argument `options` and only the [output format definition](http://math.lbl.gov/voro++/doc/custom.html) (`-c`) is lifted to an extra character argument `output_definition`. `tessellate()` returns a character vector containing the raw output of Voro++ with one vector element corresponding to one row. Depending on the structure of this raw output, different parsing functions are required to transform it to an useful R object. At the moment, ``bleiglas`` only provides one such function: `bleiglas::read_polygon_edges()`. It is configured to read data produced with the Voro++ output format string `%i*%P*%t`, which returns polygon edge coordinates. These are necessary for the default ``bleiglas`` workflow illustrated in the example below. Future versions of the package may include other parsing functions for different pipelines.
[…]
