+++ paper/paper.md
@@ -35,11 +35,11 @@ Voronoi tessellation is implemented in many R packages, maybe most prominently i
 
 # Core functionality
 
-``bleiglas`` provides the `bleiglas::tessellate()` function which is a command line utility wrapper for Voro++. It requires Voro++ to be installed locally. `tessellate()` takes input points in the form of a `data.frame` with an integer ID and three numeric coordinate columns. Additional Voro++ [options](http://math.lbl.gov/voro++/doc/cmd.html) can be set with a character argument `options` and only the [output format definition](http://math.lbl.gov/voro++/doc/custom.html) (`-c`) is lifted to an extra character argument `output_definition`. `tessellate()` returns a character vector containing the raw output of Voro++ with one vector element corresponding to one row. Depending on the structure of this raw output, different parsing functions are required to transform it to an useful R object. At the moment, ``bleiglas`` only provides one such function: `bleiglas::read_polygon_edges()`. It is configured to read data produced with the Voro++ output format string `%i*%P*%t`, which returns polygon edge coordinates. These are necessary for the default ``bleiglas`` workflow illustrated in the example below. Future versions of the package may include other parsing functions for different pipelines.
+``bleiglas`` provides the `bleiglas::tessellate()` function which is a command line utility wrapper for Voro++. It requires Voro++ to be installed locally. `tessellate()` takes input points in the form of a `data.frame` with an integer ID and three numeric coordinate columns. Additional Voro++ [options](http://math.lbl.gov/voro++/doc/cmd.html) can be set with a character argument `options` and only the [output format definition](http://math.lbl.gov/voro++/doc/custom.html) (`-c`) is lifted to an extra character argument `output_definition`. `tessellate()` returns a character vector containing the raw output of Voro++ with one vector element corresponding to one row. Depending on the structure of this raw output, different parsing functions are required to transform it to a useful R object. At the moment, ``bleiglas`` provides one such function: `bleiglas::read_polygon_edges()`. It is configured to read data produced with the Voro++ output format string `%i*%P*%t`, which returns polygon edge coordinates. These are necessary for the default ``bleiglas`` workflow illustrated in the example below. Future versions of the package may include other parsing functions for different pipelines.
 
 The output of `read_polygon_edges()` is (for performance reasons) a `data.table` [@Dowle2019] object that can be used with `bleiglas::cut_polygons()`. This function now shoulders the core task of cutting the 3D, polgyon-filled Voro++ output box into 2D slices. 3D data is notoriously difficult to plot and read. Extracting and visualizing slices is therefore indispensable. The necessary algorithm for each 3D polygon can be summarised as finding the cutting point of each polygon edge line with the requested cutting surface and then defining the convex hull of the cutting points as a result 2D polygon. We implemented the line-segment-plane-intersection operation via Rcpp [@Eddelbuettel2017] in C++ for better performance and used `grDevices::chull()` for the convex hull search. The output of `cut_polygons()` is a list (for each cut surface) of lists (for each polygon) of `data.table`s (3D coordinates for each 2D cutting point). Optionally, and in case of horizontal (z-axis) cuts with spatial coordinates on the x- and y-axis, this output can be transformed to an `sf` [@Pebesma2018] object via `bleiglas::cut_polygons_to_sf()`. This significantly simplifies subsequent map plotting.
 
[…]
