### 2021-04-08 e53ce26
replaced hyperlinks in figure caption with latex syntax
