+++ paper/paper.md
@@ -53,7 +53,7 @@ For \autoref{fig:plot_bleiglas} we cut these polygons into 2D time slices that c
 
 For the final \autoref{fig:plot_prediction_grid} we applied temporal resampling with `bleiglas::predict_grid()` to record observational error caused by radiocarbon dating uncertainty. For reasons of computational performance we kept the number of resampling runs and the spatial resolution in this example low. Nevertheless the advantages of the resampling approach are well visible: areas and periods with uncertain attribution to one burial type or the other are indicated as such.
 
-![Graves in the research area (rectangular frame) dating between 2200 and 800 calBC as extracted from the \href{radon-b.ufg.uni-kiel.de}{Radon-B database}. The classes of the variable burial type are distinguished by colour, the ones of burial construction by shape. The map projection is EPSG:102013 (Europe Albers Equal Area Conic) and the base layer data is taken from the \href{https://www.naturalearthdata.com}{Natural Earth project}.\label{fig:plot_map}](03_map_plot.jpeg)
+![Graves in the research area (rectangular frame) dating between 2200 and 800 calBC as extracted from the \protect\href{radon-b.ufg.uni-kiel.de}{Radon-B database}. The classes of the variable burial type are distinguished by colour, the ones of burial construction by shape. The map projection is EPSG:102013 (Europe Albers Equal Area Conic) and the base layer data is taken from the \protect\href{https://www.naturalearthdata.com}{Natural Earth project}.\label{fig:plot_map}](03_map_plot.jpeg)
 
 ![Graves in 3D space defined by two spatial (x and y in km) and a temporal (z in years calBC) dimension with Voronoi polygons constructed by Voro++. Each red dot represents one grave with known burial type, the fine black lines the edges of the result polygons and the rectangular wireframe box the research area now in space and time.\label{fig:plot_3D}](05_3D_plot.jpeg)
 
