### 2021-04-08 #3092 @nevrome
Release: https://github.com/nevrome/bleiglas/releases/tag/1.0.0
DOI: https://doi.org/10.17605/OSF.IO/AGMD6
