### 2021-04-09 #3092 @vissarion
> Release: https://github.com/nevrome/bleiglas/releases/tag/1.0.0
> DOI: https://doi.org/10.17605/OSF.IO/AGMD6

Thanks. I am checking the metadata of the archive and the list of authors is not the same as in the submitted paper. Could you please update?
