### 2021-04-09 #3092 @vissarion
Thanks @stschiff and @nevrome for your cooperation here!

This is an interesting use case of computational geometry to archeology and I believe the package will be of use by practitioners on that or related fields.

I would like to thank again @corybrunson and @fabian-s for the careful reviews!

I have already recommended acceptance.

At this point, some EiC will take over to make final checks and publish the paper.
