### 2021-04-09 #3092 @danielskatz
@nevrome - As AEiC on duty currently, I've proofread the paper, and have made some suggestions in https://github.com/nevrome/bleiglas/pull/20  Please merge this or let me know what you disagree with, and we can proceed to acceptance and publication.
