### 2021-04-09 #3092 @danielskatz
👋 @openjournals/dev - this is ready for final acceptance, except that there's a reference that carries into the margin and perhaps off the page.  Any idea how to fix this?
