### 2021-04-10 #3092 @arfon
@danielskatz - this PR should fix the rendering issue: https://github.com/nevrome/bleiglas/pull/21

@nevrome - could you please merge this?
