### 2021-04-10 #3092 @arfon
@corybrunson , @fabian-s – many thanks for your reviews here and to @vissarion for editing this submission. JOSS relies upon the volunteer efforts of people like you, and we simply wouldn't be able to do this without you ✨

@nevrome – your papers is now accepted and published in JOSS :zap::rocket::boom:
